# Disable stencil instancing