On my device any effect on frame times is lost in the normal variation, and the mean of only 3 runs actually came out a little faster without instancing.

```
before:
| boston     | 16.4 ms | 16.4 ms | 16.5 ms 
| paris      |  6.5 ms |  6.7 ms | 16.5 ms 
| paris2     | 16.5 ms |  6.5 ms | 16.5 ms 
| alps       | 17.6 ms | 11.5 ms | 13.3 ms 
| us east    |  7.0 ms |  6.8 ms | 16.4 ms 
| greater la | 16.5 ms | 16.5 ms | 16.5 ms 
| sf         | 11.7 ms | 12.0 ms | 15.7 ms 
| oakland    | 16.5 ms | 16.5 ms |  6.4 ms 
| germany    | 14.9 ms | 17.4 ms | 18.4 ms 
Average frame time: 13.7 ms, 12.3, 15.1 mean=13.7

after:
| boston     | 16.5 ms | 16.5 ms | 16.5 ms
| paris      |  6.8 ms |  6.5 ms |  6.6 ms
| paris2     |  6.5 ms | 16.5 ms |  6.5 ms
| alps       | 13.1 ms | 16.5 ms | 16.9 ms
| us east    |  7.0 ms | 16.5 ms | 16.5 ms
| greater la | 16.5 ms | 16.5 ms | 16.5 ms
| sf         | 16.1 ms | 11.3 ms | 17.1 ms
| oakland    | 16.5 ms |  6.5 ms |  6.4 ms
| germany    | 18.6 ms | 17.0 ms | 15.5 ms
Average frame time: 13.1 ms, 13.7, 13.2 mean=13.3
```