@stefankarschti 2023-10-18
_iPhone 6s_ results
```
Benchmark completed.
Result:
| boston     | 10.6 ms | 94.5 fps |
| paris      | 15.4 ms | 65.1 fps |
| paris2     | 17.5 ms | 57.3 fps |
| alps       | 13.1 ms | 76.5 fps |
| us east    | 14.1 ms | 71.1 fps |
| greater la | 19.4 ms | 51.6 fps |
| sf         | 25.5 ms | 39.2 fps |
| oakland    | 15.2 ms | 65.9 fps |
| germany    | 21.5 ms | 46.5 fps |
Average frame time: 16.9 ms
Average FPS: 59.2
```